[#noissue] Pass the class metadata to BasedMatcherIndex as a Supplier
BasedMatcherIndex and IndexValueMatcher depend on Supplier<InternalClassMetadata> instead of the wrapper class. ClassMetadataSuppliers creates either a constant supplier for metadata the caller already parsed or the lazy one that parses the class bytes once, which replaces ClassMetadataWrapper.
